Move labels logic to subpackage
